requirement.txt
asgiref==3.6.0
channels==3.0.4
channels-redis==4.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
six==1.16.0
django-redis==5.2.0
sockets==1.0.0
requests
public_ip